Handle undefined __file__ in REPL usage
When evaluating a buffer using inferior-lisp mode in Emacs, the code
in the buffer is send to a running Hy REPL. It looks like the REPL
doesn't have the __file__ variable defined. This breaks evaluation of
forms that use `test-set` for example, because evaluation of __file__
throws an exception. This patch verifies that __file__ actually does
exists and handles situations when it doesn't.